Update all dependency uses, move TransformerService and TransformerInterface to Transformers package
