+++ plugins/woocommerce/src/Admin/RemoteInboxNotifications/TransformerService.php
@@ -19,7 +19,7 @@ class TransformerService extends DeprecatedClassFacade {
 	 *
 	 * @var string
 	 */
-	protected static $facade_over_classname = 'Automattic\WooCommerce\Admin\RemoteSpecs\RuleProcessors\TotalPaymentsVolumeProcessor';
+	protected static $facade_over_classname = 'Automattic\WooCommerce\Admin\RemoteSpecs\RuleProcessors\Transformers\TransformerService';
 
 	/**
 	 * The version that this class was deprecated in.
