+++ plugins/woocommerce/tests/legacy/unit-tests/woocommerce-admin/remote-inbox-notifications/not-rule-processor.php
@@ -6,7 +6,7 @@
  */
 
 use Automattic\WooCommerce\Admin\RemoteInboxNotifications\NotRuleProcessor;
-use Automattic\WooCommerce\Admin\RemoteInboxNotifications\RuleEvaluator;
+use Automattic\WooCommerce\Admin\RemoteSpecs\RuleProcessors\RuleEvaluator;
 
 /**
  * class WC_Admin_Tests_RemoteInboxNotifications_NotRuleProcessor
